+++ backend/src/main/java/com/now/naaga/game/application/GameService.java
@@ -9,8 +9,8 @@ import com.now.naaga.game.domain.GameStatus;
 import com.now.naaga.game.exception.GameException;
 import com.now.naaga.game.repository.GameRepository;
 import com.now.naaga.member.application.MemberService;
+import com.now.naaga.member.application.dto.MemberCommand;
 import com.now.naaga.member.domain.Member;
-import com.now.naaga.member.dto.MemberCommand;
 import com.now.naaga.place.application.PlaceService;
 import com.now.naaga.place.domain.Place;
 import com.now.naaga.place.domain.Position;
