+++ backend/src/main/java/com/now/naaga/game/presentation/dto/GameResponse.java
@@ -1,6 +1,7 @@
 package com.now.naaga.game.presentation.dto;
 
 import com.now.naaga.game.domain.Game;
+import com.now.naaga.place.presentation.dto.PlaceResponse;
 
 public class GameResponse {
