+++ backend/src/main/java/com/now/naaga/member/application/dto/MemberCommand.java
@@ -1,4 +1,4 @@
-package com.now.naaga.member.dto;
+package com.now.naaga.member.application.dto;
 
 public class MemberCommand {
